requirements.txt
Flask==2.3.2
gunicorn==21.2.0
requests==2.31.0
pyjwt[crypto]==2.8.0